+++ backend/test/createApplication.ts
@@ -2,7 +2,7 @@ import { MikroORM } from '@mikro-orm/core';
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 
-import { AppModule } from '../src/app.module';
+import { AppModule } from '../src/AppModule';
 
 export const createApplication = async (): Promise<INestApplication> => {
 	const module = await Test.createTestingModule({
